Variables are now sorted by name before code generation. Variables in arbitrary order would lead to different hashes.
The library will now be imported in order to find out if it exists. The library will be compiled if an error occurs. Before I was searching for a file with the libraries name. Seems simpler and more robust on other platforms.

Preprocessor directive added for weave style access of array dimension length